+++ dds/src/xtypes/data_storage.rs
@@ -1,10 +1,9 @@
 use crate::xtypes::{
     dynamic_type::{DynamicData, DynamicDataFactory},
     error::{XTypesError, XTypesResult},
-    type_support::{Type, TypeSupport},
+    type_support::TypeSupport,
 };
 use alloc::{string::String, vec::Vec};
-use dust_dds::xtypes::dynamic_type::{DynamicType, ExtensibilityKind, TypeDescriptor, TypeKind};
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum DataStorage {
@@ -659,458 +658,5 @@ impl<T: TypeSupport> DataStorageMapping for T {
             DataStorage::ComplexValue(x) => Ok(T::create_sample(x)),
             _ => Err(XTypesError::InvalidType),
         }
-    }
-}
-
-impl Type for u8 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::UINT8,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-impl Type for i8 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::INT8,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for u16 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::UINT16,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for i16 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::INT16,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for u32 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::UINT32,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for i32 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::INT32,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for u64 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::UINT64,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for i64 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::INT64,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for bool {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::BOOLEAN,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for f32 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::FLOAT32,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for f64 {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::FLOAT64,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for char {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::CHAR8,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: true,
-        },
-        member_list: &[],
-    };
-}
-
-impl<T> Type for Box<T> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::ALIAS,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl<T> Type for Option<T> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::ALIAS,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: None,
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for String {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::STRING8,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: None,
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl<T: Type, const N: usize> Type for [T; N] {
-    const TYPE: DynamicType = crate::xtypes::dynamic_type::DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::ARRAY,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(N as u32),
-            element_type: Some(T::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl<T: TypeSupport> Type for Vec<T> {
-    const TYPE: DynamicType = crate::xtypes::dynamic_type::DynamicType {
-        descriptor: &crate::xtypes::dynamic_type::TypeDescriptor {
-            kind: crate::xtypes::dynamic_type::TypeKind::SEQUENCE,
-            name: "SequenceComplexValue",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(T::TYPE),
-            key_element_type: None,
-            extensibility_kind: crate::xtypes::dynamic_type::ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for Vec<u8> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::SEQUENCE,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(u8::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for Vec<u32> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::SEQUENCE,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(u32::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for Vec<i32> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::SEQUENCE,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(i32::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for Vec<u16> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::SEQUENCE,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(u16::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-impl Type for Vec<String> {
-    const TYPE: DynamicType = DynamicType {
-        descriptor: &TypeDescriptor {
-            kind: TypeKind::SEQUENCE,
-            name: "",
-            base_type: None,
-            discriminator_type: None,
-            bound: Some(u32::MAX),
-            element_type: Some(String::TYPE),
-            key_element_type: None,
-            extensibility_kind: ExtensibilityKind::Final,
-            is_nested: false,
-        },
-        member_list: &[],
-    };
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn really_create_dynamic_sample<T: TypeSupport>(v: T) -> DynamicData {
-        let mut data = crate::xtypes::dynamic_type::DynamicDataFactory::create_data(T::TYPE);
-        v.create_dynamic_sample(&mut data);
-        data
-    }
-
-    #[test]
-    fn basic_type_should_create_sample() {
-        #[derive(TypeSupport)]
-        pub struct Inner {
-            x: u16,
-        }
-
-        let mut inner = DynamicDataFactory::create_data(Inner::TYPE);
-        inner.set_uint16_value(0, 2).unwrap();
-
-        let v = Inner { x: 2 };
-        assert_eq!(really_create_dynamic_sample(v), inner);
-    }
-
-    #[test]
-    fn complex_type_should_create_sample() {
-        #[derive(TypeSupport)]
-        pub struct Inner {
-            x: u16,
-        }
-        #[derive(TypeSupport)]
-        pub struct Outer {
-            y: Inner,
-        }
-
-        let v = Outer { y: Inner { x: 2 } };
-        let mut inner = DynamicDataFactory::create_data(Inner::TYPE);
-        inner.set_uint16_value(0, 2).unwrap();
-        let mut outer = DynamicDataFactory::create_data(Outer::TYPE);
-        outer.set_complex_value(0, inner).unwrap();
-
-        assert_eq!(really_create_dynamic_sample(v), outer);
-    }
-
-    #[test]
-    fn vector_of_basic_type_should_create_sample() {
-        #[derive(TypeSupport)]
-        pub struct Inner {
-            x: Vec<u16>,
-        }
-        let mut inner = DynamicDataFactory::create_data(Inner::TYPE);
-        inner.set_uint16_values(0, vec![2]).unwrap();
-
-        let v = Inner { x: vec![2] };
-        assert_eq!(really_create_dynamic_sample(v), inner);
-    }
-
-    #[test]
-    fn vector_of_complex_type_should_create_sample() {
-        #[derive(TypeSupport)]
-        pub struct Inner {
-            x: u16,
-        }
-        #[derive(TypeSupport)]
-        pub struct Outer {
-            list: Vec<Inner>,
-        }
-
-        let v = Outer {
-            list: vec![Inner { x: 2 }],
-        };
-        let mut inner = DynamicDataFactory::create_data(Inner::TYPE);
-        inner.set_uint16_value(0, 2).unwrap();
-        let mut outer = DynamicDataFactory::create_data(Outer::TYPE);
-        outer.set_complex_values(0, vec![inner]).unwrap();
-
-        assert_eq!(really_create_dynamic_sample(v), outer);
     }
 }
